requirements.txt
# API Server
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0
python-whois>=0.8.0
builtwith>=1.3.6

# Network and Scanning
scapy>=2.5.0
python-nmap>=0.7.1
dnspython>=2.4.0
shodan>=1.30.0
censys==2.2.0

# Web Testing
selenium>=4.15.0
aiohttp>=3.9.0
websockets>=12.0

# Password and Cryptography
paramiko>=3.3.0
bcrypt>=4.1.0
cryptography>=41.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# CLI and Output
colorama>=0.4.6
tabulate>=0.9.0
rich>=13.7.0
click>=8.1.7
pyfiglet>=1.0.2

# Async and Performance
asyncio>=3.4.3
aiofiles>=23.2.1
concurrent-log-handler>=0.9.25

# Utilities
python-dotenv>=1.0.0
validators>=0.22.0
tldextract>=5.1.0
fake-useragent>=1.4.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Optional (uncomment if needed)
# selenium-wire>=5.1.0
# playwright>=1.40.0
